refactor: some search methods moved to base plugin
